requirements.txt
python-docx>=0.8.11
google-genai>=0.3.0
python-dotenv>=1.0.0
